fixed issue with textfields getting reset when users type in input
what happened was that for some reason, the GettingStarted page had functioning TextFields, but the other pages didn't when the Project data type was implemented, with the text of the textfield being reset as the user left it, likely due to the text of the controller being set at the beginning of the build function.

It was fixed by moving the declerations up and into the contstuctor of the class to only have it run once.